+++ eth_defi/event_reader/progress_update.py
@@ -12,6 +12,7 @@ import datetime
 
 from tqdm_loggable.auto import tqdm
 
+from eth_defi.compat import native_datetime_utc_fromtimestamp
 from eth_defi.event_reader.reader import ProgressUpdate
 
 
@@ -102,7 +103,7 @@ class TQDMProgressUpdate(ProgressUpdate):
             self.progress_bar = self.create_progress_bar(start_block, end_block)
 
         if last_timestamp:
-            friendly_time = datetime.datetime.utcfromtimestamp(last_timestamp).strftime("%Y-%m-%d %H:%M")
+            friendly_time = native_datetime_utc_fromtimestamp(last_timestamp).strftime("%Y-%m-%d %H:%M")
         else:
             friendly_time = "NA"