+++ tests/enzyme/test_guard_enzyme_uniswap_v3.py
@@ -5,7 +5,6 @@
 - Check some negative cases for unauthroised transactions
 """
 
-import datetime
 import random
 
 import flaky
@@ -13,15 +12,10 @@ import pytest
 from eth_account import Account
 from eth_account.signers.local import LocalAccount
 from eth_typing import HexAddress
-from eth_defi.terms_of_service.acceptance_message import (
-    generate_acceptance_message,
-    get_signing_hash,
-    sign_terms_of_service,
-)
 from web3 import Web3
 from web3.contract import Contract
-from web3.middleware import construct_sign_and_send_raw_middleware
 
+from eth_defi.compat import construct_sign_and_send_raw_middleware, native_datetime_utc_now
 from eth_defi.deploy import deploy_contract
 from eth_defi.enzyme.deployment import EnzymeDeployment, RateAsset
 from eth_defi.enzyme.generic_adapter_vault import deploy_vault_with_generic_adapter
@@ -29,6 +23,11 @@ from eth_defi.enzyme.uniswap_v3 import prepare_swap
 from eth_defi.enzyme.vault import Vault
 from eth_defi.hotwallet import HotWallet
 from eth_defi.middleware import construct_sign_and_send_raw_middleware_anvil
+from eth_defi.terms_of_service.acceptance_message import (
+    generate_acceptance_message,
+    get_signing_hash,
+    sign_terms_of_service,
+)
 from eth_defi.token import TokenDetails
 from eth_defi.trace import (
     TransactionAssertionError,
@@ -86,7 +85,7 @@ def acceptance_message(web3: Web3) -> str:
     # Generate the message user needs to sign in their wallet
     signing_content = generate_acceptance_message(
         1,
-        datetime.datetime.utcnow(),
+        native_datetime_utc_now(),
         "https://example.com/terms-of-service",
         random.randbytes(32),
     )
@@ -340,6 +339,7 @@ def test_enzyme_guarded_trade_singlehop_uniswap_v3(
         pool_fees=[POOL_FEE_RAW],
         token_in_amount=200 * 10**6,  # 200 USD
     )
+    print(f"*****{prepared_tx=}*****")
 
     tx_hash = prepared_tx.transact({"from": asset_manager, "gas": 1_000_000})
     assert_transaction_success_with_explanation(web3, tx_hash)